feat: abstract ChatRepository interface and create ObjectBoxChatRepository implementation
