+++ lib/data/chat_repository.dart
@@ -1,8 +1,7 @@
-import '../generated/objectbox/objectbox.g.dart';
 import 'chat_message.dart';
 
 abstract interface class ChatRepository {
-  /// Returns up to [_maxMessages] messages ordered newest-first.
+  /// Returns up to 100 messages ordered newest-first.
   /// If [since] is provided, only messages with `timestamp > since` are returned.
   List<ChatMessage> loadMessages({int? since});
 
@@ -17,88 +16,4 @@ abstract interface class ChatRepository {
   Future<List<ChatMessage>> searchMessages(String query);
 
   void dispose();
-}
-
-class ChatRepo implements ChatRepository {
-  static const int _maxMessages = 100;
-
-  late final Store _store;
-  late final Box<ChatMessage> _box;
-
-  ChatRepo._();
-
-  static Future<ChatRepository> create() async {
-    final repo = ChatRepo._();
-    repo._store = await openStore();
-    repo._box = repo._store.box<ChatMessage>();
-
-    return repo;
-  }
-
-  @override
-  void dispose() => _store.close();
-
-  @override
-  void clearAll() => _box.removeAll();
-
-  /// Returns up to [_maxMessages] messages ordered newest-first (descending timestamp).
-  @override
-  List<ChatMessage> loadMessages({int? since}) {
-    final builder = since != null
-        ? _box.query(ChatMessage_.timestamp.greaterThan(since))
-        : _box.query();
-
-    final query =
-        (builder..order(ChatMessage_.timestamp, flags: Order.descending))
-            .build()
-          ..limit = _maxMessages;
-    try {
-      return query.find();
-    } finally {
-      query.close();
-    }
-  }
-
-  @override
-  Future<List<ChatMessage>> searchMessages(String query) async {
-    if (query.isEmpty) return [];
-    final q = _box
-        .query(ChatMessage_.content.contains(query, caseSensitive: false))
-        .order(ChatMessage_.timestamp, flags: Order.descending)
-        .build()
-      ..limit = _maxMessages;
-    try {
-      return await q.findAsync();
-    } finally {
-      q.close();
-    }
-  }
-
-  @override
-  void saveMessage({
-    required ChatMessageRoleEnum role,
-    required String content,
-  }) {
-    _box.put(
-      ChatMessage(
-        content: content,
-        timestamp: DateTime.now().millisecondsSinceEpoch,
-        role: role.value,
-      ),
-    );
-    _trimToLimit();
-  }
-
-  void _trimToLimit() {
-    final count = _box.count();
-    final excess = count - _maxMessages;
-    if (excess <= 0) return;
-    final query = _box.query().order(ChatMessage_.timestamp).build()
-      ..limit = excess;
-    try {
-      _box.removeMany(query.findIds());
-    } finally {
-      query.close();
-    }
-  }
 }
